separacion de componentes

+++ Controllers/CategoriaController.cs
@@ -1,7 +1,10 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using FluentValidation;
+using FluentValidation.Results;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ReactVentas.Models;
+using System.Linq.Expressions;
 
 namespace ReactVentas.Controllers
 {
@@ -9,11 +12,14 @@ namespace ReactVentas.Controllers
     [ApiController]
     public class CategoriaController : ControllerBase
     {
+        private IValidator<Categoria> _validator;
         private readonly DBREACT_VENTAContext _context;
-        public CategoriaController(DBREACT_VENTAContext context)
+        private ValidationResult validationResult;
+        public CategoriaController(IValidator<Categoria> validator, DBREACT_VENTAContext context)
         {
             _context = context;
-
+            _validator = validator;
+            validationResult = new();
         }
         [HttpGet]
         [Route("Lista")]
@@ -22,10 +28,10 @@ namespace ReactVentas.Controllers
             List<Categoria> lista = new List<Categoria>();
             try
             {
-                lista = await _context.Categoria.OrderByDescending(c => c.IdCategoria).ToListAsync();
+                lista = await _context.Categoria.OrderByDescending(c => c.IdCategoria).AsNoTracking().ToListAsync();
                 return StatusCode(StatusCodes.Status200OK, lista);
             }
-            catch (Exception ex)
+            catch
             {
                 return StatusCode(StatusCodes.Status500InternalServerError, lista);
             }
@@ -36,13 +42,18 @@ namespace ReactVentas.Controllers
         public async Task<IActionResult> Guardar([FromBody] Categoria request) {
             try
             {
-                await _context.Categoria.AddAsync(request);
-                await _context.SaveChangesAsync();
+                validationResult =  await _validator.ValidateAsync(request);
+                if (validationResult.IsValid)
+                {
+                    await _context.Categoria.AddAsync(request);
+                    await _context.SaveChangesAsync();
+                    return StatusCode(StatusCodes.Status200OK, "ok");
+                }
+                return StatusCode(StatusCodes.Status500InternalServerError);
 
-                return StatusCode(StatusCodes.Status200OK, "ok");
             }
-            catch(Exception ex) {
-                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            catch {
+                return StatusCode(StatusCodes.Status500InternalServerError);
             }
         }
 
@@ -56,8 +67,8 @@ namespace ReactVentas.Controllers
 
                 return StatusCode(StatusCodes.Status200OK, "ok");
             }
-            catch (Exception ex) {
-                return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+            catch {
+                return StatusCode(StatusCodes.Status500InternalServerError);
             }
         }
 
@@ -66,10 +77,14 @@ namespace ReactVentas.Controllers
         public async Task<IActionResult> Eliminar(int id) {
             try
             {
-                Categoria categoria = _context.Categoria.Find(id);
-                _context.Categoria.Remove(categoria);
-                await _context.SaveChangesAsync();
-                return StatusCode(StatusCodes.Status200OK, "ok");
+                Categoria ?categoria = _context.Categoria.Find(id);
+                if (categoria is not null)
+                {
+                    _context.Categoria.Remove(categoria);
+                    await _context.SaveChangesAsync();
+                    return StatusCode(StatusCodes.Status200OK, "ok");
+                }
+                return StatusCode(StatusCodes.Status204NoContent);
             }
             catch (Exception ex) {
                 return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
